etl/requirements.txt
functions-framework==3.*
requests
pandas
google-cloud-bigquery
pyarrow
